requirements.txt
pillow>=5.4.1
python-twitter>=3.5
pyyaml>=3.13
requests>=2.21.0
tqdm>=4.31.1
urllib3>=1.24.1
